Move currently deprecated milkmen tests
